Adding optional text wrapping support
Text wrapping will occur if the string is bigger than the maximum length (basically, before the text starts resizing too much) and only if the input string has no newlines. If it has, we'll assume the mod author wants to format it themselves, and place it as is.

So this change won't affect anyone already using newlines in their tooltips, but will allow them to opt in.